move method so it is easier to reuse
