+++ packages/types/src/index.ts
@@ -23,6 +23,7 @@ export * from "./tool.js"
 export * from "./type-fu.js"
 export * from "./vscode.js"
 export * from "./kilocode.js" // kilocode_change
+export * from "./kilocode-utils.js" // kilocode_change
 export * from "./usage-tracker.js" // kilocode_change
 
 export * from "./providers/index.js"
