+++ packages/types/src/kilocode-utils.ts
@@ -1,0 +1,17 @@
+export function getKiloBaseUriFromToken(kilocodeToken?: string) {
+	if (kilocodeToken) {
+		try {
+			const payload_string = kilocodeToken.split(".")[1]
+			if (!payload_string) return "https://api.kilocode.ai"
+
+			const payload_json =
+				typeof atob !== "undefined" ? atob(payload_string) : Buffer.from(payload_string, "base64").toString()
+			const payload = JSON.parse(payload_json)
+			//note: this is UNTRUSTED, so we need to make sure we're OK with this being manipulated by an attacker; e.g. we should not read uri's from the JWT directly.
+			if (payload.env === "development") return "http://localhost:3000"
+		} catch (_error) {
+			console.warn("Failed to get base URL from Kilo Code token")
+		}
+	}
+	return "https://api.kilocode.ai"
+}
